Replace _dynamo.config with an dataclass object
Summary:

Currently _dynamo.config (and friends) is written as module but
is an object with its data backed by a dictionary. This change
makes it a dataclass instead.

Nested namespace via class (`class cpp:` etc in _inductor/config.py)
are also becoming dataclass.

For now, the top level dataclass generated dynamically from the module
attr declarations. Next step is to also remove that and just use a
normal dataclass.

The goal of having dataclasses for config is that there is a name for
the config class, and therefore we can write functions that take those
configs as input.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

reset test files